+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The changelog for `SuperwallKit`. Also see the [releases](https://github.com/superwall/Superwall-iOS/releases) on GitHub.
 
+## 4.12.5
+
+### Enhancements
+
+- Adds microphone permission request support.
+
+### Fixes
+
+- Fixes issue where the notification permission prompt would not appear if provisional notification permission was already granted.
+
 ## 4.12.4
 
 ### Enhancements